Refactor Attributes in the Inspector and update documentation
