+++ src/client/src/modules/inspector/components/header/InspectorHeader.styled.ts
@@ -1,12 +1,12 @@
 import styled from "styled-components";
 import { Color } from "../../../../assets/color/Color";
 
-interface InspectorHeaderContainerProps {
+interface InspectorHeaderBoxProps {
   top?: number;
   color: string;
 }
 
-export const InspectorHeaderContainer = styled.div<InspectorHeaderContainerProps>`
+export const InspectorHeaderBox = styled.div<InspectorHeaderBoxProps>`
   display: flex;
   color: ${Color.BLACK};
   height: 44px;
@@ -15,13 +15,3 @@ export const InspectorHeaderContainer = styled.div<InspectorHeaderContainerProps
   padding-top: ${(props) => (props.top ? props.top : 0)}px;
   background-color: ${(props) => props.color}!important;
 `;
-
-export const InspectorHeaderNodeInfo = styled.div`
-  display: flex;
-  gap: 10px;
-  flex-shrink: 0;
-  margin: 14px 0 10px 10px;
-  font-weight: bold;
-`;
-
-export const InspectorHeaderNodeInfoText = styled.span``;
