+++ src/client/src/modules/inspector/components/tabs/InspectorTabsComponent.styled.ts
@@ -1,0 +1,11 @@
+import styled from "styled-components";
+
+export const InspectorHeaderNodeInfo = styled.div`
+  display: flex;
+  gap: 10px;
+  flex-shrink: 0;
+  margin: 14px 0 10px 10px;
+  font-weight: bold;
+`;
+
+export const InspectorHeaderNodeInfoText = styled.span``;
