+++ src/client/src/modules/inspector/components/tabs/helpers/index.ts
@@ -3,3 +3,4 @@ export * from "./GetInspectorTabsColor";
 export * from "./GetTabId";
 export * from "./GetInspectorText";
 export * from "./IsRelationEdge";
+export * from "./GetInspectorHeaderText";
